migrate compat into a separate folder and separate tf and torch import
