+++ merlin/core/compat/__init__.py
@@ -15,10 +15,9 @@
 #
 
 # pylint: disable=unused-import
-import os
 import warnings
 
-from packaging import version
+from merlin.core.has_gpu import HAS_GPU  # noqa pylint: disable=unused-import
 
 try:
     import psutil
@@ -29,8 +28,6 @@ try:
     from numba import cuda
 except ImportError:
     cuda = None
-
-from merlin.core.has_gpu import HAS_GPU
 
 
 def pynvml_mem_size(kind="total", index=0):
@@ -137,77 +134,3 @@ try:
     import dask_cudf
 except ImportError:
     dask_cudf = None
-
-try:
-    import tensorflow
-
-    def configure_tensorflow(memory_allocation=None, device=None):
-        """Utility to help configure tensorflow to not use 100% of gpu memory as buffer"""
-        tf = tensorflow
-        total_gpu_mem_mb = device_mem_size(kind="total", cpu=(not HAS_GPU)) / (1024**2)
-
-        if memory_allocation is None:
-            memory_allocation = os.environ.get("TF_MEMORY_ALLOCATION", 0.5)
-
-        if float(memory_allocation) < 1:
-            memory_allocation = total_gpu_mem_mb * float(memory_allocation)
-        memory_allocation = int(memory_allocation)
-        assert memory_allocation < total_gpu_mem_mb
-
-        if HAS_GPU:
-            tf_devices = tf.config.list_physical_devices("GPU")
-
-            if len(tf_devices) == 0:
-                raise ImportError("TensorFlow is not configured for GPU")
-
-            for tf_device in tf_devices:
-                try:
-                    tf.config.set_logical_device_configuration(
-                        tf_device,
-                        [tf.config.LogicalDeviceConfiguration(memory_limit=memory_allocation)],
-                    )
-                except RuntimeError:
-                    warnings.warn(
-                        "TensorFlow runtime already initialized, may not be enough memory for cudf"
-                    )
-                try:
-                    tf.config.experimental.set_virtual_device_configuration(
-                        tf_device,
-                        [
-                            tf.config.experimental.VirtualDeviceConfiguration(
-                                memory_limit=memory_allocation
-                            )
-                        ],
-                    )
-                except RuntimeError as e:
-                    # Virtual devices must be set before GPUs have been initialized
-                    warnings.warn(str(e))
-
-        # versions using TF earlier than 2.3.0 need to use extension
-        # library for dlpack support to avoid memory leak issue
-        __TF_DLPACK_STABLE_VERSION = "2.3.0"
-        if version.parse(tf.__version__) < version.parse(__TF_DLPACK_STABLE_VERSION):
-            try:
-                from tfdlpack import from_dlpack
-            except ModuleNotFoundError as e:
-                message = (
-                    "If using TensorFlow < 2.3.0, you must install tfdlpack-gpu extension library"
-                )
-                raise ModuleNotFoundError(message) from e
-
-        else:
-            from tensorflow.experimental.dlpack import from_dlpack
-
-        return from_dlpack
-
-    configure_tensorflow()
-
-    from tensorflow.python.framework import ops as tf_ops
-except ImportError:
-    tensorflow = None
-    tf_ops = None
-
-try:
-    import torch
-except ImportError:
-    torch = None
